Rename RateLimitEnvelope and move to rate_limits.rs
